+++ native/spark-expr/src/agg_funcs/mod.rs
@@ -1,7 +1,15 @@
 mod avg;
 mod avg_decimal;
 mod correlation;
+mod covariance;
+mod stddev;
+mod variance;
+mod sum_decimal;
 
 pub use avg::Avg;
 pub use avg_decimal::AvgDecimal;
 pub use correlation::Correlation;
+pub use covariance::Covariance;
+pub use stddev::Stddev;
+pub use variance::Variance;
+pub use sum_decimal::SumDecimal;
