+++ native/spark-expr/src/array_funcs/mod.rs
@@ -1,0 +1,7 @@
+mod list_extract;
+mod get_array_struct_fields;
+mod array_insert;
+
+pub use list_extract::ListExtract;
+pub use get_array_struct_fields::GetArrayStructFields;
+pub use array_insert::ArrayInsert;
