+++ native/spark-expr/src/conditional_funcs/mod.rs
@@ -1,0 +1,3 @@
+mod if_expr;
+
+pub use if_expr::IfExpr;
