+++ native/spark-expr/src/math_funcs/mod.rs
@@ -1,0 +1,3 @@
+mod negative;
+
+pub use negative::{NegativeExpr, create_negate_expr};
