+++ native/spark-expr/src/predicate_funcs/mod.rs
@@ -1,0 +1,3 @@
+mod rlike;
+
+pub use rlike::RLike;
